FRESTYLE-369 refactor: 楽観ロック群 2 リポジトリを GORM から sqlc へ完全移行する（段 4） (#2313)
* FRESTYLE-369 test: course_chapters の読み書き（一覧の doc 除外・並び順・Update の列選択と updated_at 更新・削除）を結合テストで固定する

GORM→sqlc 移行の前後で振る舞いが同一であることを保証する土台。楽観ロックの 409/404 は
既存テストで固定済みのため、それ以外の CRUD を実 Postgres で追加で固定する。


* FRESTYLE-369 chore: rich_documents の sqlc クエリを追加し生成する

読み書き・楽観ロック更新・論理削除・一覧（doc 本体を除く軽量サマリ）を生 SQL で定義し
make sqlc の生成物を同梱する。repository の差し替えは次コミット。


* FRESTYLE-369 refactor: rich_document_repository を GORM から sqlc へ完全移行する

読み書き・論理削除・一覧をすべて sqlc 生成コードへ差し替え、GORM のクエリ呼び出しを 0 にする
（接続プールは *gorm.DB.DB() から借りるだけ。wiring は段 5）。楽観ロック更新は RETURNING の
:one で更新後の行を受け、0 行（sql.ErrNoRows）を存在確認で 404 / 409 に切り分ける契約を保つ
（GORM の RowsAffected + 存在確認と同じ）。uuid PK・NULL 可 company_id・class22→400 翻訳・
deleted_at IS NULL の論理削除除外を維持する。


* FRESTYLE-369 chore: course_chapters の sqlc クエリを追加し生成する

読み 4（会社一覧・コース一覧・単体・件数集計）と書き 5（作成・メタ更新・本文の楽観ロック更新・
削除・コース単位削除）を生 SQL で定義し make sqlc の生成物を同梱する。コース内一覧は本文
（doc・jsonb）を返さない軽量列だけの SELECT にする。repository の差し替えは次コミット。


* FRESTYLE-369 refactor: teaching_material_repository を GORM から sqlc へ完全移行する

読み 4・書き 5（raw SQL 4 本を含む）をすべて sqlc 生成コードへ差し替え、GORM のクエリ呼び出しを
0 にする（接続プールは *gorm.DB.DB() から借りるだけ。gorm.ErrRecordNotFound は 404 シグナルとして
残す。wiring は段 5）。本文の楽観ロック更新は RETURNING の :one で更新後の行を受け、0 行
（sql.ErrNoRows）を存在確認で 404 / 409 に切り分ける契約を保つ。コース内一覧は本文 doc を返さない
軽量列だけの SELECT にし、Update は now() で updated_at を進める GORM の Updates(map) 挙動を維持する。


* fix: 章の Update で 0 行時の sql.ErrNoRows を握り潰さず not-found として返す

UpdateChapter は :one + RETURNING のため対象行が無いと sql.ErrNoRows を返すが、
それをそのまま外へ返していたため handler が解釈できず 500 になっていた。
gorm.ErrRecordNotFound へ翻訳して 404 に落とす。

移行前の GORM 版は Updates(map) で 0 行でも nil（無言の成功）だったが、
それは「消えた行を更新したことにする」silent failure なので戻さない。
呼び出し側は先に GetByID しているので、0 行になるのは取得と更新の間に
行が消された競合のみで、そのとき正しい答えは 404。

存在し得ない id（uint64 が int64 に収まらない）も同じ理由で not-found に揃えた。


* test: 存在しない id への Update / Delete の契約を結合テストで固定する

Update / Delete に「存在しない id」のケースが無かったため、0 行更新が無言で
成功する欠陥が結合テストでも変異テストでも検出できていなかった。

- Update: 存在しない id・取得後に消えた章の 2 経路で gorm.ErrRecordNotFound を固定
  （巻き添えで他の行が書き換わらないことも同時に確認する）
- Delete / DeleteByCourse: 0 行でも nil（冪等）であることを固定。GORM 版・sqlc 版とも
  0 行で nil を返すことを実測済みで、Update と違い失われた書き込みが無いため契約を変えない


* fix: 範囲外 id の書き込みが「何も書かずに成功」を返すのをやめる

bigint に収まらない uint64 を渡された書き込み系が nil を返しており、
1 行も書いていないのに呼び出し側は作成・更新できたと誤認していた
（usecase はそのまま 201 / 200 を返し、次の取得で初めて「無い」と分かる）。

rich_documents の Create が同ファイル内の uuid.Parse 失敗と揃っていなかったのが
発端だが、grep すると同じ silent-nil が書き込み系 18 か所にあったため一括で直す。
errOutOfRangeID / outOfRangeIDError を ids.go に置き、諦めた列名を添えて返す。

読み取り系の「存在し得ない id = 0 件 / not found」は書き込みを伴わず
呼び出し側の解釈も正しいため変更しない。


* test: 書き込み系が範囲外 id を成功として返さないことを結合テストで固定する

読み取り側の「0 件でも null を返さない」表駆動テストと対になる書き込み側の表。
書き込み系 18 経路それぞれで bigint に収まらない id を渡し、エラーが返ることを
固定する（silent-nil に戻すと 18 件すべてが落ちることを確認済み）。


* fix: bigint 列へのパラメータを ::int で 32bit に絞らず ::bigint に揃える

rich_documents / course_chapters の revision・schema_version・sort_order は
いずれも bigint（information_schema で実測）。SQL 側が ::int（int4）で
キャストしていたため sqlc も int32 パラメータを生成し、int32 を超える値は
エラーも出ないまま負数へ巻き戻って保存されていた（実測: 2147483648 → -2147483648）。

::bigint に直して再生成し、呼び出し側の int32() 変換も int64() に揃える。

courses.sort_order の ::int は schema.sql の宣言（integer）と一致しているため
ここでは触らない（宣言と実型の食い違いは別途）。


* test: int32 を超える revision / schema_version / sort_order の往復を固定する

bigint 列への書き込みが 32bit に切り詰められると、エラーを出さずに負数へ
巻き戻った別の値が保存される。::int に戻すと 2147483648 が -2147483648 として
保存され両テストが落ちることを確認済み。


* style: 1 行に収めていた WHERE 句を改行して他のクエリと揃える

DELETE / UPDATE の WHERE が 1 行に収まっており、他のクエリ（WHERE を改行）と
不揃いだった。queries/*.sql 全体を確認し、文の WHERE が同一行にある 4 か所
（course_chapter 2・course 1・admin_invitation 1）を揃える。

FILTER (WHERE ...) / 式の中の副問い合わせ / 部分インデックスの述語は
同じ構文ではないので触っていない。生成物も再生成して同梱する。


* perf: 会社別の章一覧で使われない本文 jsonb を読み出さない

domain.TeachingMaterial.Doc は json:"-" で応答に出ないため、会社別一覧の
戻り値の Doc は usecase / handler / テストのどこからも読まれていない
（.Doc の参照を全 Go ファイルで確認済み）。それでも全章の doc(jsonb) を
読み出して転送していた。

コース別一覧と同じ軽量な列構成に揃える。行型は sqlc が SELECT ごとに
別々に生成するため、company 別の行を course 別の行型へ変換してから
domain へ写す（列構成がずれたらコンパイルエラーになる）。


* test: 会社別の章一覧が本文を返さないことを結合テストで固定する

doc を入れた章を含めても一覧の Doc が nil のままであることを確認する。
doc 更新は updated_at を now() へ進めるため、並び順の固定より前に行う。


* fix: courses.sort_order を実型どおり bigint として扱い 32bit へ絞らない

sqlc の型付けスキーマが courses.sort_order を integer と宣言していたが、
本番の実型は bigint（information_schema で確認）。この食い違いにより
INSERT のパラメータが int4 に落ち、int32 を超える値はエラーも出さずに
負数へ巻き戻って保存されていた（2147483648 → -2147483648 を実測）。

course_chapters / rich_documents と同じ欠陥で、それらは同 PR 内で修正済み。
宣言を実型に合わせ、キャストを ::bigint に、呼び出し側を int64 に揃える。


* test: courses.sort_order が int32 を超える値を切り詰めないことを固定する

修正前の状態（宣言 integer + ::int + int32）を再現すると
expected 2147483648 / actual -2147483648 で落ちることを確認済み。


---------

+++ backend/internal/adapter/persistence/queries/course.sql
@@ -24,7 +24,7 @@ VALUES (
   sqlc.arg(description),
   sqlc.arg(category),
   sqlc.arg(language),
-  COALESCE(NULLIF(sqlc.arg(sort_order)::int, 0), 100),
+  COALESCE(NULLIF(sqlc.arg(sort_order)::bigint, 0), 100),
   sqlc.arg(is_published),
   sqlc.arg(created_at),
   sqlc.arg(updated_at)
@@ -46,4 +46,5 @@ RETURNING updated_at;
 
 -- name: DeleteCourse :exec
 -- コースを物理削除する（courses は soft delete 列を持たない）。
-DELETE FROM courses WHERE id = sqlc.arg(id);
+DELETE FROM courses
+WHERE id = sqlc.arg(id);
